[PM-4060] split build script into target specific modules
