+++ apps/desktop/desktop_native/target_build/mod.rs
@@ -1,0 +1,5 @@
+#[cfg_attr(target_os = "linux", path = "unix.rs")]
+#[cfg_attr(target_os = "windows", path = "windows.rs")]
+#[cfg_attr(target_os = "macos", path = "macos.rs")]
+mod target_build;
+pub use target_build::*;
