+++ apps/desktop/desktop_native/target_build/windows.rs
@@ -1,0 +1,1 @@
+pub fn target_specific_build() {}
